api/requirements.txt
flask-cors==3.0.10
flask-restx==0.5.0
flask-marshmallow==0.14.0
flask-sqlalchemy==2.5.1
marshmallow-sqlalchemy==0.26.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
